// PP-CoE-CodeApp/src/generated/models/PowerPlatformforAdminsV2Model.ts
export interface GetConnectorByIdResponse {
  name?: string;
  id?: string;
  type?: string;
  properties?: {
    displayName?: string;
    iconUri?: string;
    iconBrandColor?: string;
    apiEnvironment?: string;
    isCustomApi?: boolean;
    blobUrisAreProxied?: boolean;
    runtimeUrls?: string[];
    primaryRuntimeUrl?: string;
    doNotUseApiHubNetRuntimeUrl?: string;
    metadata?: {
      source?: string;
      brandColor?: string;
      allowSharing?: boolean;
      useNewApimVersion?: string;
      version?: {
        previous?: string;
        current?: string;
      };
    };
    capabilities?: string[];
    interfaces?: Record<string, unknown>;
    description?: string;
    createdTime?: string;
    changedTime?: string;
    releaseTag?: string;
    tier?: string;
    publisher?: string;
    rateLimit?: number;
    apiVersion?: string;
  };
}
// ...
export interface ListConnectorsResponse {
  value?: GetConnectorByIdResponse[];
}